+++ src/demos/elements/DemoTag.vue
@@ -1,0 +1,14 @@
+<template>
+	<bulma-box>
+		<BulmaTag>Some text</BulmaTag>
+	</bulma-box>
+</template>
+
+<style scoped>
+
+</style>
+
+<script setup lang="ts">
+
+	import BulmaBox from "../../vbComponents/elements/BulmaBox.vue";
+	import BulmaTag from "../../vbComponents/elements/BulmaTag.vue";</script>