+++ src/vuebulma.ts
@@ -18,6 +18,7 @@ import BulmaImage from "./vbComponents/elements/BulmaImage.vue";
 import BulmaNotification from "./vbComponents/elements/BulmaNotification.vue";
 import BulmaProgressbar from "./vbComponents/elements/BulmaProgressbar.vue";
 import BulmaTable from "./vbComponents/elements/BulmaTable.vue";
+import BulmaTag from "./vbComponents/elements/BulmaTag.vue";
 import BulmaFileInput from "./vbComponents/form/BulmaFileInput.vue";
 import BulmaInput from "./vbComponents/form/BulmaInput.vue";
 import BulmaSelect from "./vbComponents/form/BulmaSelect.vue";
@@ -27,10 +28,11 @@ import BulmaFormControlGroup from "./vbComponents/form/groups/BulmaFormControlGr
 import BulmaFormField from "./vbComponents/form/groups/BulmaFormField.vue";
 import BulmaButtonGroup from "./vbComponents/layouts/BulmaButtonGroup.vue";
 import BulmaHero from "./vbComponents/layouts/BulmaHero.vue";
+import BulmaTagGroup from "./vbComponents/layouts/BulmaTagGroup.vue";
+import BulmaTile from "./vbComponents/layouts/BulmaTile.vue";
 import BulmaColumn from "./vbComponents/layouts/Columns/BulmaColumn.vue";
 import BulmaColumnGroup from "./vbComponents/layouts/Columns/BulmaColumnGroup.vue";
 import BulmaLevel from "./vbComponents/layouts/Level/BulmaLevel.vue";
-import BulmaTile from "./vbComponents/layouts/Tile/BulmaTile.vue";
 
 const vuebulma = {
 	install(Vue: any) {
@@ -64,8 +66,8 @@ const vuebulma = {
 		Vue.component("vbColumnGroup", BulmaColumnGroup);
 		Vue.component("vbColumn", BulmaColumn);
 		Vue.component("vbTile", BulmaTile);
-
-
+		Vue.component("vbTag", BulmaTag);
+		Vue.component("vbTagGroup", BulmaTagGroup);
 	}
 };
 
@@ -103,5 +105,7 @@ export {
 	BulmaColumnGroup as vbColumnGroup,
 	BulmaColumn as vbColumn,
 	BulmaTile as vbTile,
+	BulmaTagGroup as vbTagGroup,
+	BulmaTag as vbTag,
 	vuebulma
 };